+++ gix-macros/tests/macros.rs
@@ -1,0 +1,1 @@
+mod momo;
